feat(coding-agent): priority to commit model for titles with fallback to smol
